requirements_agent.txt
python-socketio[client]==5.11.2
psutil==5.9.8
requests==2.31.0
websocket-client==1.8.0
mss==9.0.1
Pillow>=10.4.0
